Refactor: move UI functions from Building to BuildingDescriptions
